requirements.txt
torch>=2.2
torchaudio>=2.2
whisperx>=3.1.1
pyannote.audio>=3.1
tqdm>=4.66
pyyaml>=6.0
python-dotenv>=1.0
jiwer>=3.0
